style(MathematicalExpressionParser) : 더 명확한 이름을 위해 클래스와 인터페이스에 Long타입을 명시
