--- src/main/java/calculator/parser/MathematicalExpressionParser.java
@@ -1,6 +1,0 @@
-package calculator.parser;
-
-public interface MathematicalExpressionParser {
-
-	public long[] parse(String text, String separatorRegex);
-}
